# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BedemWAF Control API
  version: 0.1.0
  description: REST management API for BedemWAF tenants, apps, policies, and audit event references.
servers:
  - url: http://localhost:8081
security:
  - AdminBearerAuth: []
paths:
  /healthz:
    get:
      summary: Liveness check
      security: []
      responses:
        "200":
          description: Service is alive
  /readyz:
    get:
      summary: Readiness check
      security: []
      responses:
        "200":
          description: Database is reachable
        "503":
          $ref: "#/components/responses/Error"
  /v1/tenants:
    get:
      summary: List tenants
      responses:
        "200":
          description: Tenant list
    post:
      summary: Create tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantRequest"
      responses:
        "201":
          description: Tenant created
        "400":
          $ref: "#/components/responses/Error"
  /v1/apps:
    get:
      summary: List apps
      responses:
        "200":
          description: App list
    post:
      summary: Create app with primary origin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppRequest"
      responses:
        "201":
          description: App created
        "400":
          $ref: "#/components/responses/Error"
  /v1/apps/{app_id}:
    get:
      summary: Get app
      parameters:
        - $ref: "#/components/parameters/AppID"
      responses:
        "200":
          description: App
        "404":
          $ref: "#/components/responses/Error"
    patch:
      summary: Update app
      parameters:
        - $ref: "#/components/parameters/AppID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppRequest"
      responses:
        "200":
          description: App updated
        "400":
          $ref: "#/components/responses/Error"
  /v1/apps/{app_id}/policies:
    get:
      summary: List policies for app
      parameters:
        - $ref: "#/components/parameters/AppID"
      responses:
        "200":
          description: Policy list
    post:
      summary: Create policy draft
      parameters:
        - $ref: "#/components/parameters/AppID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyRequest"
      responses:
        "201":
          description: Policy created
  /v1/apps/{app_id}/active-policy:
    get:
      summary: Get active compiled policy for an app
      parameters:
        - $ref: "#/components/parameters/AppID"
      responses:
        "200":
          description: Gateway-ready policy JSON
        "404":
          $ref: "#/components/responses/Error"
  /v1/policies/{policy_id}:
    get:
      summary: Get policy
      parameters:
        - $ref: "#/components/parameters/PolicyID"
      responses:
        "200":
          description: Policy
        "404":
          $ref: "#/components/responses/Error"
    patch:
      summary: Update mutable policy draft with optimistic locking
      parameters:
        - $ref: "#/components/parameters/PolicyID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePolicyRequest"
      responses:
        "200":
          description: Policy draft updated
        "409":
          $ref: "#/components/responses/Error"
  /v1/policies/{policy_id}/publish:
    post:
      summary: Publish policy draft to immutable policy version
      parameters:
        - $ref: "#/components/parameters/PolicyID"
      responses:
        "201":
          description: Policy version published
  /v1/gateway/apps/{hostname}/policy:
    get:
      summary: Get active gateway policy by hostname
      security:
        - GatewayBearerAuth: []
      parameters:
        - name: hostname
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Gateway-ready policy JSON
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayPolicy"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /v1/events:
    get:
      summary: Search WAF audit events
      parameters:
        - name: tenant_id
          in: query
          schema:
            type: string
        - name: app_id
          in: query
          schema:
            type: string
        - name: host
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
        - name: client_ip
          in: query
          schema:
            type: string
        - name: matched_rule_id
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        "200":
          description: WAF audit events
        "400":
          $ref: "#/components/responses/Error"
  /v1/events/{request_id}:
    get:
      summary: Get WAF audit event by request ID
      parameters:
        - name: request_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: WAF audit event
        "404":
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    AdminBearerAuth:
      type: http
      scheme: bearer
      description: Static MVP admin token from BEDEMWAF_ADMIN_API_KEY.
    GatewayBearerAuth:
      type: http
      scheme: bearer
      description: Static MVP gateway token from BEDEMWAF_GATEWAY_API_KEY.
  parameters:
    AppID:
      name: app_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PolicyID:
      name: policy_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: JSON error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
    CreateTenantRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{0,62}$"
        metadata:
          type: object
    CreateAppRequest:
      type: object
      required: [tenant_id, name, slug, hostnames, origin_url]
      properties:
        tenant_id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        hostnames:
          type: array
          minItems: 1
          items:
            type: string
        origin_url:
          type: string
          format: uri
          description: http or https origin URL without query or fragment.
        metadata:
          type: object
    UpdateAppRequest:
      type: object
      properties:
        name:
          type: string
        hostnames:
          type: array
          items:
            type: string
        origin_url:
          type: string
          format: uri
        status:
          type: string
          enum: [active, disabled]
        metadata:
          type: object
    CreatePolicyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        mode:
          type: string
          enum: [count, block]
          default: count
        snapshot:
          type: object
          description: MVP policy JSON snapshot. Actions may be allow, count, block, or rate_limit.
    UpdatePolicyRequest:
      type: object
      required: [expected_updated_at]
      properties:
        name:
          type: string
        mode:
          type: string
          enum: [count, block]
        enabled:
          type: boolean
        expected_updated_at:
          type: string
          format: date-time
          description: Last observed policy updated_at value.
        snapshot:
          type: object
    GatewayPolicy:
      type: object
      required:
        - tenant_id
        - app_id
        - policy_id
        - policy_version_id
        - mode
        - origin
        - ip_sets
        - custom_rules
        - rate_limits
        - waf
      properties:
        tenant_id:
          type: string
        app_id:
          type: string
        policy_id:
          type: string
        policy_version_id:
          type: string
        mode:
          type: string
          enum: [count, block]
        origin:
          type: object
        ip_sets:
          type: object
        custom_rules:
          type: array
        rate_limits:
          type: array
        waf:
          type: object